Added a parameter redis_config_abs_dir_env in OP attribute which is the name of the environment variable that stores the path of the Redis configuration file to be read by the specified model for the operation of one machine and multiple models.

+++ tensorflow_recommenders_addons/dynamic_embedding/python/ops/redis_table_ops.py
@@ -21,6 +21,8 @@ from __future__ import print_function
 import functools
 import json
 import os
+from re import T
+import warnings
 
 from tensorflow.python.eager import context
 from tensorflow.python.framework import dtypes
@@ -140,29 +142,67 @@ class RedisTable(LookupInterface):
     self._embedding_name = (self._name.split('_mht_', 1))[0]
     self._config = config
 
-    with open(self._config.redis_config_abs_dir, 'r', encoding='utf-8') as f0:
-      params_load = json.load(f0)
-      self._redis_params = self.default_redis_params.copy()
-      for k in self._redis_params.keys():
-        if k in params_load:
-          self._redis_params[k] = params_load[k]
+    self.redis_config_file_exist = False
+    self.redis_config_file_create = False
 
-    with open(self._config.redis_config_abs_dir, 'w', encoding='utf-8') as f1:
-      f1.write(json.dumps(self._redis_params, indent=2, ensure_ascii=True))
+    if self._config.redis_config_abs_dir_env:
+      if self._config.redis_config_abs_dir_env in os.environ:
+        self._config.redis_config_abs_dir = os.getenv(
+            self._config.redis_config_abs_dir_env)
+      else:
+        raise ValueError(
+            "Config redis_config_abs_dir_env in RedisTableConfig is not None, but can not find the "
+            + self._config.redis_config_abs_dir_env +
+            " in system environment variable.")
+      self.redis_config_file_exist = os.path.exists(
+          self._config.redis_config_abs_dir)
+      if self.redis_config_file_exist == False:
+        raise ValueError(
+            "Config redis_config_abs_dir_env in RedisTableConfig is not None, but the FILE which path stored in environment variable "
+            + self._config.redis_config_abs_dir_env + " DOES NOT EXIST.")
+    elif self._config.redis_config_abs_dir_env is None and "TFRA_REDIS_CONFIG_PATH" in os.environ:
+      self._config.redis_config_abs_dir = os.getenv("TFRA_REDIS_CONFIG_PATH")
+      self.redis_config_file_exist = os.path.exists(
+          self._config.redis_config_abs_dir)
+      warnings.warn(
+          "TFRA-Redis try to use environment variable TFRA_REDIS_CONFIG_PATH regardless redis_config_abs_dir in RedisTableConfig."
+      )
+      if self.redis_config_file_exist == False:
+        raise ValueError(
+            "environment variable TFRA_REDIS_CONFIG_PATH exists, but the FILE which path stored in TFRA_REDIS_CONFIG_PATH DOES NOT EXIST. Please create a FILE in the corresponding path or delete the environment variable TFRA_REDIS_CONFIG_PATH."
+        )
+    elif self._config.redis_config_abs_dir_env is None and "TFRA_REDIS_CONFIG_PATH" not in os.environ and self._config.redis_config_abs_dir:
+      self.redis_config_file_exist = os.path.exists(
+          self._config.redis_config_abs_dir)
+      if self.redis_config_file_exist == False:
+        raise ValueError(
+            "Config redis_config_abs_dir in RedisTableConfig is not None and redis_config_abs_dir_env is None, but the FILE "
+            + self._config.redis_config_abs_dir +
+            " which path is redis_config_abs_dir DOES NOT EXIST.")
+    elif self._config.redis_config_abs_dir_env is None and "TFRA_REDIS_CONFIG_PATH" not in os.environ and self._config.redis_config_abs_dir is None:
+      self.redis_config_file_create = True
+      self._config.redis_config_abs_dir = "/tmp/tmp_TFRA_Redis_config_file.json"
+      warnings.warn(
+          "Both redis_config_abs_dir_env and redis_config_abs_dir in RedisTableConfig are None, now creating a temporary config file in /tmp/tmp_TFRA_Redis_config_file.json."
+      )
+    else:
+      raise ValueError(
+          "TFRA-Redis didn't get the correct RedisTableConfig class initial parameter."
+      )
 
-    if self._config.redis_config_abs_dir_env in os.environ and os.getenv(
-        self._config.redis_config_abs_dir_env):
-      with open(os.getenv(self._config.redis_config_abs_dir_env),
-                'r',
+    if self.redis_config_file_create == True and self.redis_config_file_exist == False:
+      with open(self._config.redis_config_abs_dir, 'w+',
                 encoding='utf-8') as f0:
+        f0.write(
+            json.dumps(self.default_redis_params, indent=2, ensure_ascii=True))
+    else:
+      with open(self._config.redis_config_abs_dir, 'r', encoding='utf-8') as f0:
         params_load = json.load(f0)
         self._redis_params = self.default_redis_params.copy()
         for k in self._redis_params.keys():
           if k in params_load:
             self._redis_params[k] = params_load[k]
-      with open(os.getenv(self._config.redis_config_abs_dir_env),
-                'w',
-                encoding='utf-8') as f1:
+      with open(self._config.redis_config_abs_dir, 'w', encoding='utf-8') as f1:
         f1.write(json.dumps(self._redis_params, indent=2, ensure_ascii=True))
 
     self._shared_name = None
